Smarter logic of workflow handling when a replay is done.
In the past, the workflow snapshot was always re-created,
and it had to always be an archive. The logic has been
improved, so when the path to the snapshot does exist,
it is checked whether it is a directory. In any case,
the snapshot is not overwritten.

This is needed to ease the VRE workflow development loop,
so a directory can be used for the workflow.

Also, `workflow_archive` special output is honoured, so
it is not passed down to the nextflow execution.