Updated openvre-api-tool to v0.5.2, and also updated the code.
The workflow execution code was relying on an old version of openvre-api-tool
library, so it was not able to receive with the output metadata. The code
now receives it properly, although it is not being used yet inside.

Also, fixed the issue that nextflow version was properly detected in
workflows which contain the needed directive inside nextflow.config, but
it was not used later to launch the execution of the workflow.